pyproject.toml
[tool.poetry]
name = "whatsappsms"
version = "0.1.0"
description = ""
authors = [" GenyaElamkov < GenyaElamkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
whatsapp-api-client-python = "^0.0.34"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
